DGN: use int instead of long (https://trac.osgeo.org/gdal/ticket/5252). Code paths not actually triggered by the OGR DGN driver
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@26505 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrsf_frmts/dgn/dgnwrite.cpp
@@ -1073,7 +1073,7 @@ DGNCreateArcElem( DGNHandle hDGN, int nType,
     }
     else
     {
-        memcpy( psArc->quat, panQuaternion, sizeof(long)*4 );
+        memcpy( psArc->quat, panQuaternion, sizeof(int)*4 );
     }
 
 /* -------------------------------------------------------------------- */
@@ -1294,7 +1294,7 @@ DGNCreateConeElem( DGNHandle hDGN,
     memset( psCone->quat, 0, sizeof(int) * 4 );
     if( panQuaternion != NULL )
     {
-        memcpy( psCone->quat, panQuaternion, sizeof(long)*4 );
+        memcpy( psCone->quat, panQuaternion, sizeof(int)*4 );
     }
     else {
       psCone->quat[0] = 1 << 31;
